requirements.txt
python-telegram-bot==20.7
feedparser==6.0.10
requests==2.31.0
pytz==2023.3
